implemented MLP classification

+++ gv_experiments/train_MLP_classifier_partition_split_fprints.py
@@ -18,65 +18,24 @@ import json
 from experiments.pl_experiment import Classifier_Experiment
 from data_split.data_utils import DataSplitter
 from models.data_loaders import DrugResistanceDataset_Fingerprints, SampleEmbDataset
-from models.classifier import AMR_Classifier
+from models.classifier import MLP_Classifier
+import sys
 
 
-
-
-if __name__=="__main__":
-
-    parser = ArgumentParser()
-
-    parser.add_argument("--experiment_name", type=str, default="test1")
-    parser.add_argument("--experiment_group", type=str, default="noLeak_split_fprint")
-    parser.add_argument("--seed", type=int, default=0)
-
-    parser.add_argument("--driams_dataset", type=str, choices=['A', 'B', 'C', 'D'], default="B")
-    parser.add_argument("--driams_long_table", type=str,
-                        default="../processed_data/DRIAMS_combined_long_table.csv")
-    parser.add_argument("--spectra_matrix", type=str,
-                        default="../data/DRIAMS-B/spectra_binned_6000_2018.npy")
-    parser.add_argument("--drugs_df", type=str,
-                        default="../processed_data/drug_fingerprints.csv")
-
-
-    parser.add_argument("--fingerprint_class", type=str, default="MACCS", choices=["MACCS", "morgan_512", "morgan_1024", "pubchem"])
-    parser.add_argument("--fingerprint_size", type=int, default=167)
-
-    parser.add_argument("--drug_emb_type", type=str, default="fingerprint", choices=["fingerprint", "vae_embedding"])
-    parser.add_argument("--drug_hidden_layers", type=int, default=2)
-    parser.add_argument("--drug_hidden_dim", type=int, default=64)
-
-    parser.add_argument("--species_embedding_dim", type=int, default=16)
-    parser.add_argument("--drug_embedding_dim", type=int, default=16)
-    parser.add_argument(
-        "--sample_hidden_layers", type=int, default=2
-    )
-    parser.add_argument("--sample_embedding_dim", type=int, default=16)
-
-    parser.add_argument("--conv_out_size", type=int, default=64)
-
-    parser.add_argument("--n_epochs", type=int, default=50)
-    parser.add_argument("--batch_size", type=int, default=128)
-    parser.add_argument("--patience", type=int, default=10)
-    parser.add_argument("--learning_rate", type=float, default=0.001)
-    parser.add_argument("--weight_decay", type=float, default=1e-5)
-    parser.add_argument("--num_workers", type=int, default=1)
-
-    parser.add_argument("--random_state", type=int, default=42)
-
-    args = parser.parse_args()
+def main(args):
     config = vars(args)
 
-    output_folder = join("outputs", args.experiment_group, args.experiment_name+"_"+str(args.seed))
+    output_folder = join("outputs", args.experiment_group, args.experiment_name, str(args.seed))
     if not exists(output_folder):
         os.makedirs(output_folder)
 
+    metrics_folder = join("outputs", args.experiment_group, args.experiment_name, "metrics")
+    if not exists(metrics_folder) and args.seed==0:
+        os.makedirs(metrics_folder)
+
+    experiment_folder = join("outputs", args.experiment_group, args.experiment_name)
     # if exists(join(output_folder, "embeddings", "drugs_embeddings.csv")):
     #     sys.exit(0)
-    metrics_folder = join("outputs", args.experiment_group, args.experiment_name+"_metrics")
-    if not exists(metrics_folder):
-        os.makedirs(metrics_folder)
 
     driams_long_table = pd.read_csv(args.driams_long_table)
     spectra_matrix = np.load(args.spectra_matrix)
@@ -87,7 +46,8 @@ if __name__=="__main__":
     trainval_df, test_df = dsplit.combination_train_test_split(dsplit.long_table, test_size=0.2, random_state=args.seed)
     # train_df, val_df = dsplit.combination_train_test_split(trainval_df, test_size=0.2, random_state=args.seed)
     train_df, val_df = dsplit.baseline_train_test_split(trainval_df, test_size=0.2, random_state=args.seed)
-
+    test_df.to_csv(join(output_folder, "test_set.csv"), index=False)
+    # sys.exit(0)
 
     train_dset = DrugResistanceDataset_Fingerprints(train_df, spectra_matrix, drugs_df, species_list, fingerprint_class=config["fingerprint_class"])
     print(train_dset[0])
@@ -103,8 +63,11 @@ if __name__=="__main__":
 
 
     # Save configuration
-    with open(join(output_folder, "config.json"), "w") as f:
-        json.dump(config, f)
+    if not exists(join(experiment_folder, "config.json")):
+        del config["seed"]
+        with open(join(experiment_folder, "config.json"), "w") as f:
+            json.dump(config, f)
+
 
 
     train_loader = DataLoader(
@@ -114,11 +77,12 @@ if __name__=="__main__":
     test_loader = DataLoader(
         test_dset, batch_size=args.batch_size, shuffle=False, num_workers=args.num_workers)
 
-    model = AMR_Classifier(config)
+    model = MLP_Classifier(config)
     experiment = Classifier_Experiment(config, model)
 
-    with open(join(output_folder, "architecture.txt"), "w") as f:
-        f.write(experiment.model.__repr__())
+    if not exists(join(experiment_folder, "architecture.txt")):
+        with open(join(experiment_folder, "architecture.txt"), "w") as f:
+            f.write(experiment.model.__repr__())
 
 
 
@@ -147,17 +111,45 @@ if __name__=="__main__":
     with open(join(metrics_folder, "test_metrics_{}.json".format(args.seed)), "w") as f:
         json.dump(test_results[0], f, indent=2)
 
-    # print("Saving embeddings..")
-    # experiment.model.eval()
-    # embeddings_folder = join(output_folder, "embeddings")
-    # if not exists(embeddings_folder):
-    #     os.makedirs(embeddings_folder)
 
-    # # Species learned embeddings
-    # species_list = [idx2species[i] for i in range(len(idx2species))]
-    # species_lrn_embs = pd.DataFrame(experiment.model.species_emd.weight.detach().numpy(), 
-    #                 index=species_list, columns=["z{}".format(i) for i in range(config["species_embedding_dim"])])
-    # species_lrn_embs.index.name = "species"
-    # species_lrn_embs.to_csv(join(embeddings_folder, "species_learned_embeddings.csv"))
-    
     print("Testing complete")
+
+
+
+if __name__=="__main__":
+
+    parser = ArgumentParser()
+
+    parser.add_argument("--experiment_name", type=str, default="MLP_class")
+    parser.add_argument("--experiment_group", type=str, default="MLP_random_split_fprint")
+    parser.add_argument("--seed", type=int, default=0)
+
+    parser.add_argument("--driams_dataset", type=str, choices=['A', 'B', 'C', 'D'], default="B")
+    parser.add_argument("--driams_long_table", type=str,
+                        default="../processed_data/DRIAMS_combined_long_table.csv")
+    parser.add_argument("--spectra_matrix", type=str,
+                        default="../data/DRIAMS-B/spectra_binned_6000_2018.npy")
+    parser.add_argument("--drugs_df", type=str,
+                        default="../processed_data/drug_fingerprints.csv")
+
+
+    parser.add_argument("--fingerprint_class", type=str, default="all", choices=["all", "MACCS", "morgan_512", "morgan_1024", "pubchem"])
+    # parser.add_argument("--fingerprint_size", type=int, default=167)
+
+    parser.add_argument("--n_hidden_layers", type=int, default=5)
+    parser.add_argument("--hidden_size", type=int, default=1024)
+    parser.add_argument("--input_size", type=int, default=8089)
+
+
+    parser.add_argument("--n_epochs", type=int, default=50)
+    parser.add_argument("--batch_size", type=int, default=128)
+    parser.add_argument("--patience", type=int, default=10)
+    parser.add_argument("--learning_rate", type=float, default=0.001)
+    parser.add_argument("--weight_decay", type=float, default=1e-5)
+    parser.add_argument("--num_workers", type=int, default=0)
+
+    parser.add_argument("--random_state", type=int, default=42)
+
+    args = parser.parse_args()
+    args.num_workers = os.cpu_count()
+    main(args)
